refactor(control-plane): split control plane tests
